Rename battleEventChannel file to battleSaga
