+++ packages/shared/src/game/match/combat/index.ts
@@ -1,1 +1,1 @@
-export { battleSaga, startBattle, BattleEvent, BattleFinishEvent, BattleTurnEvent, BATTLE_FINISH_EVENT, BATTLE_TURN_EVENT } from "./battleEventChannel";
+export { battleSaga, startBattle, BattleEvent, BattleFinishEvent, BattleTurnEvent, BATTLE_FINISH_EVENT, BATTLE_TURN_EVENT } from "./battleSaga";
