+++ packages/shared/src/game/match/index.ts
@@ -1,3 +1,3 @@
 export { Match } from "./match";
 
-export { battleSaga, startBattle, BATTLE_FINISH_EVENT } from "./combat/battleEventChannel";
+export { battleSaga, startBattle, BATTLE_FINISH_EVENT } from "./combat/battleSaga";
